+++ src/components/common/sharedComponents/EditableTextFieldWithClickableIcon.js
