+++ src/components/panel-body/UEpage/atomicInfoTile.js
@@ -1,6 +1,6 @@
 import Divider from '@material-ui/core/Divider';
-import LabelledBoolean from 'components/common/niceComponents/LabelledBoolean';
-import LabelledText from 'components/common/niceComponents/LabelledText';
+import LabelledBoolean from 'components/common/sharedComponents/LabelledBoolean';
+import LabelledText from 'components/common/sharedComponents/LabelledText';
 import Paper from '@material-ui/core/Paper';
 import PropTypes from 'prop-types';
 import React from 'react';
