+++ src/components/panel-body/UEpage/forms/addressForm.js
@@ -3,10 +3,10 @@ import React, { useContext, useState } from 'react';
 import D from 'i18n';
 import DialogActions from '@material-ui/core/DialogActions';
 import { Divider } from '@material-ui/core';
-import { EditableBooleanField } from 'components/common/niceComponents/EditableBooleanField';
-import { EditableTextField } from 'components/common/niceComponents/EditableTextField';
-import GenericTile from 'components/common/niceComponents/GenericTile';
-import IconButton from 'components/common/niceComponents/IconButton';
+import { EditableBooleanField } from 'components/common/sharedComponents/EditableBooleanField';
+import { EditableTextField } from 'components/common/sharedComponents/EditableTextField';
+import GenericTile from 'components/common/sharedComponents/GenericTile';
+import IconButton from 'components/common/sharedComponents/IconButton';
 import MaterialIcons from 'utils/icons/materialIcons';
 import PropTypes from 'prop-types';
 import SurveyUnitContext from '../UEContext';
