+++ src/components/panel-body/UEpage/forms/userForm.js
@@ -6,11 +6,11 @@ import D from 'i18n';
 import DateFnsUtils from '@date-io/date-fns';
 import DialogActions from '@material-ui/core/DialogActions';
 import { Divider } from '@material-ui/core';
-import { EditableTextField } from 'components/common/niceComponents/EditableTextField';
-import { EditableTextFieldWithClickableIcon } from 'components/common/niceComponents/EditableTextFieldWithClickableIcon';
-import GenericTile from 'components/common/niceComponents/GenericTile';
-import IconButton from 'components/common/niceComponents/IconButton';
-import { LabelledSwitch } from 'components/common/niceComponents/LabelledSwitch';
+import { EditableTextField } from 'components/common/sharedComponents/EditableTextField';
+import { EditableTextFieldWithClickableIcon } from 'components/common/sharedComponents/EditableTextFieldWithClickableIcon';
+import GenericTile from 'components/common/sharedComponents/GenericTile';
+import IconButton from 'components/common/sharedComponents/IconButton';
+import { LabelledSwitch } from 'components/common/sharedComponents/LabelledSwitch';
 import MaterialIcons from 'utils/icons/materialIcons';
 import PropTypes from 'prop-types';
 import SurveyUnitContext from '../UEContext';
